requirements.txt
openai
python-dotenv
PyMuPDF
python-docx
fastapi
uvicorn
python-multipart